# docs/openapi/evidsight-v1.yaml
openapi: 3.1.0
info:
  title: EvidSight 外部 API v1
  version: 1.0.0
  description: |
    EvidSight Knowledge 与 Research 服务供 Web 消费的外部 v1 契约。本文件是 Auth、
    Knowledge Base、Document、Chat、Conversation、Research、Evidence 与 Report 请求、响应
    和 SSE data 字段与约束的唯一权威来源；API.md 定义行为、
    权限、状态码与兼容语义，本文只定义 Schema。新增或修改字段必须先修订本文件，
    再经 FastAPI 路由清单 ↔ OpenAPI 路径清单 ↔ Provider 契约测试三方一致校验
    （API.md §3.1 / ROADMAP M2 退出门禁）。
servers:
  - url: ""
paths:
  /api/v1/knowledge-bases:
    post:
      summary: 创建知识库
      description: 创建知识库并返回完整对象。同名冲突返回 409 E1002。
      tags: [Knowledge Base v1]
      operationId: createKnowledgeBaseV1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseCreate"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      summary: 知识库可见集合列表
      description: |
        以单一分页端点返回当前用户可见的知识库集合，支持「全部 / 我创建的 / 组织公开」
        范围筛选与名称搜索；同一知识库在可见范围并集中只出现一次。普通用户
        scope=all 为 mine ∪ public，admin 的 all 返回全部知识库（治理可见）。
      tags: [Knowledge Base v1]
      operationId: listKnowledgeBasesV1
      parameters:
        - $ref: "#/components/parameters/ScopeParam"
        - $ref: "#/components/parameters/QParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: 可见知识库分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBaseList"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/knowledge-bases/{kb_id}:
    parameters:
      - $ref: "#/components/parameters/KbIdParam"
    get:
      summary: 知识库详情
      description: public KB 所有登录用户可读；private KB 仅 owner 或 admin 可读。
      tags: [Knowledge Base v1]
      operationId: getKnowledgeBaseV1
      responses:
        "200":
          description: 知识库详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: 更新知识库元数据
      description: 部分更新（name/description/visibility 均可选）。owner 可修改自己的 KB，admin 可修正任意 KB。
      tags: [Knowledge Base v1]
      operationId: updateKnowledgeBaseV1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseUpdate"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      summary: 删除知识库
      description: 标记 status=deleting 并分发异步清理任务；返回 204 无正文。
      tags: [Knowledge Base v1]
      operationId: deleteKnowledgeBaseV1
      responses:
        "204":
          description: 删除任务已提交，无正文
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/auth/register:
    post:
      tags: [Auth v1]
      operationId: registerV1
      summary: 注册用户
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RegisterRequest"}
      responses:
        "201":
          description: 用户已创建
          content: {application/json: {schema: {$ref: "#/components/schemas/UserSummary"}}}
        "422": {$ref: "#/components/responses/ValidationError"}
  /api/v1/auth/login:
    post:
      tags: [Auth v1]
      operationId: loginV1
      summary: 登录
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginRequest"}
      responses:
        "200":
          description: 登录成功；Refresh Token 只通过 HttpOnly Cookie 返回
          content: {application/json: {schema: {$ref: "#/components/schemas/LoginResponse"}}}
        "401": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /api/v1/auth/refresh:
    post:
      tags: [Auth v1]
      operationId: refreshV1
      summary: 轮换 Refresh Family
      responses:
        "200":
          description: Access Token 已刷新
          content: {application/json: {schema: {$ref: "#/components/schemas/RefreshResponse"}}}
        "401": {$ref: "#/components/responses/Forbidden"}
  /api/v1/auth/logout:
    post:
      tags: [Auth v1]
      operationId: logoutV1
      summary: 退出并撤销 Refresh Family
      responses:
        "204": {description: 已退出，无正文}
        "401": {$ref: "#/components/responses/Forbidden"}
  /api/v1/auth/me:
    get:
      tags: [Auth v1]
      operationId: getCurrentUserV1
      summary: 当前用户摘要
      responses:
        "200":
          description: 当前用户
          content: {application/json: {schema: {$ref: "#/components/schemas/UserSummary"}}}
        "401": {$ref: "#/components/responses/Forbidden"}
  /api/v1/auth/password:
    put:
      tags: [Auth v1]
      operationId: changePasswordV1
      summary: 修改密码并撤销全部 Refresh Family
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ChangePasswordRequest"}
      responses:
        "204": {description: 密码已修改，无正文}
        "401": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /api/v1/chat/stream:
    post:
      tags: [Chat v1]
      operationId: streamChatV1
      summary: 发起单知识库问答
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ChatStreamRequest"}
      responses:
        "200":
          description: canonical Chat SSE
          content:
            text/event-stream:
              schema: {type: string}
              x-sse-data-schemas:
                meta: {$ref: "#/components/schemas/ChatMetaEventData"}
                message.delta: {$ref: "#/components/schemas/ChatDeltaEventData"}
                sources: {$ref: "#/components/schemas/ChatSourcesEventData"}
                error: {$ref: "#/components/schemas/StreamErrorEventData"}
                done: {$ref: "#/components/schemas/ChatDoneEventData"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /api/v1/chat/generations/{generation_id}/cancel:
    post:
      tags: [Chat v1]
      operationId: cancelChatGenerationV1
      summary: 幂等取消生成
      parameters:
        - {name: generation_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "202":
          description: 取消已接受
          content: {application/json: {schema: {$ref: "#/components/schemas/ChatCancelResponse"}}}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: 页码，从 1 开始
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      required: false
      description: 每页条数
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ScopeParam:
      name: scope
      in: query
      required: false
      description: 可见范围筛选：all=可见并集（默认），mine=我创建的，public=组织公开
      schema:
        type: string
        enum: [all, mine, public]
        default: all
    QParam:
      name: q
      in: query
      required: false
      description: 按名称模糊搜索
      schema:
        type: string
        maxLength: 128
    KbIdParam:
      name: kb_id
      in: path
      required: true
      description: 知识库稳定 UUID
      schema:
        type: string
        format: uuid
...
  responses:
    Forbidden:
      description: 无权限执行此操作（E5005）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: 资源不存在或不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: 状态或名称冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationError:
      description: 请求参数校验失败（E9003）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      description: API.md §4 标准错误信封
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
      additionalProperties: false
    Error:
      type: object
      required: [error_code, message, request_id, retryable, details]
      properties:
        error_code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]+$"
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
          description: 安全、结构化、可操作信息；不含正文、密码、Token、SQL、路径或堆栈
      additionalProperties: false
    # ---- Knowledge Base ----
    KnowledgeBase:
      type: object
      required: [uuid, name, owner, visibility, status, doc_count, chunk_count, created_at]
      properties:
        uuid:
          type: string
          format: uuid
          description: 知识库稳定 UUID
        name:
          type: string
          minLength: 2
          maxLength: 128
        description:
          type: string
          maxLength: 2000
          nullable: true
        owner:
          type: string
          format: uuid
          description: owner 用户 Platform User UUID（非内部 users.id）
        visibility:
          type: string
          enum: [private, public]
        status:
          type: string
          enum: [active, deleting]
        doc_count:
          type: integer
          minimum: 0
        chunk_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    KnowledgeBaseCreate:
      type: object
      description: 创建知识库请求
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 128
          description: 名称不能为纯数字或纯空格
        description:
          type: string
          maxLength: 2000
          nullable: true
        visibility:
          type: string
          enum: [private, public]
          default: private
    KnowledgeBaseUpdate:
      type: object
      description: 更新知识库请求（部分更新，至少一个字段）
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 128
          nullable: true
        description:
          type: string
          maxLength: 2000
          nullable: true
        visibility:
          type: string
          enum: [private, public]
          nullable: true
    KnowledgeBaseList:
      type: object
      required: [total, page, page_size, items]
      properties:
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        items:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeBase"
    # ---- Document ----
...
    RegisterRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username: {type: string, minLength: 2, maxLength: 64}
        password: {type: string, minLength: 6, maxLength: 128, writeOnly: true}
    LoginRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username: {type: string, minLength: 2, maxLength: 64}
        password: {type: string, minLength: 6, maxLength: 128, writeOnly: true}
    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      additionalProperties: false
      properties:
        old_password: {type: string, minLength: 6, maxLength: 128, writeOnly: true}
        new_password: {type: string, minLength: 6, maxLength: 128, writeOnly: true}
    UserSummary:
      type: object
      required: [id, username, role, status]
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        username: {type: string}
        role: {type: string, enum: [user, admin]}
        status: {type: string, enum: [active, disabled]}
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      additionalProperties: false
      properties:
        access_token: {type: string, writeOnly: true}
        token_type: {type: string, const: bearer}
        expires_in: {type: integer, minimum: 1}
        user: {$ref: "#/components/schemas/UserSummary"}
    RefreshResponse:
      type: object
      required: [access_token, token_type, expires_in]
      additionalProperties: false
      properties:
        access_token: {type: string, writeOnly: true}
        token_type: {type: string, const: bearer}
        expires_in: {type: integer, minimum: 1}
    # ---- Chat SSE ----
    ChatStreamRequest:
      type: object
      required: [knowledge_base_id, question]
      additionalProperties: false
      properties:
        conversation_id: {type: [string, "null"], format: uuid}
        knowledge_base_id: {type: string, format: uuid}
        question: {type: string, minLength: 1, maxLength: 2000}
        deep_thinking: {type: boolean, default: false}
    ChatMetaEventData:
      type: object
      required: [generation_id, conversation_id]
      properties:
        generation_id: {type: string, format: uuid}
        conversation_id: {type: string, format: uuid}
    ChatDeltaEventData:
      type: object
      required: [delta]
      additionalProperties: false
      properties:
        delta: {type: string}
    ChatSource:
      type: object
      required: [chunk_index, doc_name, score, document_uuid, segment_id]
      properties:
        chunk_index: {type: integer, minimum: 1}
        doc_name: {type: string}
        score: {type: number}
        document_uuid: {type: string, format: uuid}
        segment_id: {type: string, format: uuid}
        page: {type: [integer, "null"], minimum: 1}
        section_title: {type: [string, "null"]}
        section_path: {type: [string, "null"]}
        preview_text: {type: [string, "null"]}
        preview_range: {$ref: "#/components/schemas/PreviewRange"}
        highlight_start: {type: [integer, "null"], minimum: 0}
        highlight_end: {type: [integer, "null"], minimum: 0}
    PreviewRange:
      type: [object, "null"]
      required: [start, end]
      properties:
        start: {type: integer, minimum: 0}
        end: {type: integer, minimum: 0}
    ChatSourcesEventData:
      type: object
      required: [chunks]
      properties:
        chunks: {type: array, items: {$ref: "#/components/schemas/ChatSource"}}
        confidence: {type: [string, "null"]}
        confidence_note: {type: [string, "null"]}
    StreamErrorEventData:
      type: object
      required: [error_code, message, retryable]
      properties:
        error_code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
    ChatDoneEventData:
      type: object
      required: [message_id, token_usage]
      properties:
        message_id: {type: integer}
        title: {type: [string, "null"]}
        token_usage:
          type: object
          required: [prompt, completion, total]
          properties:
            prompt: {type: integer, minimum: 0}
            completion: {type: integer, minimum: 0}
            total: {type: integer, minimum: 0}
    ChatCancelResponse:
      type: object
      required: [generation_id, status, idempotent_replayed]
      properties:
        generation_id: {type: string, format: uuid}
        status: {type: string, enum: [canceled]}
        idempotent_replayed: {type: boolean}
    # ---- Research ----
...